Add UserNavigation component
